requirements.txt
numpy>=1.26.0
opencv-python>=4.10.0.84
scikit-image>=0.24.0
scikit-learn>=1.5.0
pandas>=2.2.0
matplotlib>=3.8.0
tqdm>=4.66.0
pyyaml>=6.0.1
jsonschema>=4.22.0
# For future model-based explanations (optional):
# torch>=2.2.0
# captum>=0.7.0
